refactor: フックと非フックの配置違反を解消
- src/pages/useAIChatDetailLifecycle(.test) → src/hooks/（pages にフックを置かない）
- src/hooks/aiChatEmptyTree, runAIChatAction(.test) → src/lib/aiChat/（フックでないヘルパー）
- src/hooks/workflowRunOutcomeHandlers → src/lib/workflow/（同上）

AGENTS.md の配置規則に従い、import は @/ エイリアスに更新。
