+++ src/hooks/useAIChatDetailLifecycle.test.tsx
