+++ src/hooks/useAIChatDetailLifecycle.ts
